requirements.txt
python-telegram-bot[job-queue]==22.5
python-dotenv==1.2.1
pytz
